+++ examples/search/select_metrics.py
@@ -1,0 +1,10 @@
+from aim.sdk import select_metrics
+
+
+if __name__ == '__main__':
+    metrics = select_metrics((
+        'loss if '
+        '   run.archived is False '
+        '   and run.loss.train > 0.01 '
+    ))
+    print(metrics)