+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.13 Sep 21 2020
+- Add ability to query runs by metric value (gorarakelyan)
+- Add ability to query runs via SDK (gorarakelyan)
+
 ## 2.0.12 Sep 12 2020
 - Update Session to handle exceptions gracefully (gorarakelyan)